# SIMD-ize DenseHll::mergeWith