--- a/velox/common/hyperloglog/benchmarks/CMakeLists.txt
+++ b/velox/common/hyperloglog/benchmarks/CMakeLists.txt
@@ -11,18 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-add_library(velox_common_hyperloglog BiasCorrection.cpp DenseHll.cpp
-                                     SparseHll.cpp)
 
-target_link_libraries(
-  velox_common_hyperloglog
-  PUBLIC velox_memory
-  PRIVATE velox_exception)
+add_executable(velox_common_hyperloglog_dense_hll_bm DenseHll.cpp)
 
-if(${VELOX_BUILD_TESTING})
-  add_subdirectory(tests)
-endif()
-
-if(${VELOX_ENABLE_BENCHMARKS})
-  add_subdirectory(benchmarks)
-endif()
+target_link_libraries(velox_common_hyperloglog_dense_hll_bm
+                      velox_common_hyperloglog ${FOLLY_BENCHMARK})